requirements.txt
streamlit==1.22.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.81.0
google-cloud-bigquery==3.9.0
huggingface-hub==0.16.4
xlsxwriter==3.1.2
numpy==1.23.5
pandas==1.5.3
